requirements_pip.txt
flask==2.0.1
flask_restful=0.3.8
mysql-connector-python==8.0.29
